[ZVOOQ-7245] Sink pip version to github version of payture sdk
