--- payture/payture/MANIFEST.in
@@ -1,6 +1,0 @@
-# Include the license file
-include LICENSE.txt
-
-# Include the data files
-recursive-include base *
-recursive-include encodingtypes *
