--- payture/payture/payture.pyproj
@@ -1,63 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" ToolsVersion="4.0">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>e896c0f2-0653-4e9d-9bd9-254878a2f4f2</ProjectGuid>
-    <ProjectHome>.</ProjectHome>
-    <StartupFile>testapp\starttest.py</StartupFile>
-    <SearchPath>payturetypes\;testapp\</SearchPath>
-    <WorkingDirectory>.</WorkingDirectory>
-    <OutputPath>.</OutputPath>
-    <Name>payture</Name>
-    <RootNamespace>payture</RootNamespace>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
-    <DebugSymbols>true</DebugSymbols>
-    <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-    <DebugSymbols>true</DebugSymbols>
-    <EnableUnmanagedDebugging>false</EnableUnmanagedDebugging>
-  </PropertyGroup>
-  <ItemGroup>
-    <Folder Include="testapp\" />
-    <Folder Include="payturetypes\" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="payturetypes\api.py" />
-    <Compile Include="payturetypes\cardinfo.py" />
-    <Compile Include="payturetypes\constants.py" />
-    <Compile Include="payturetypes\digitalwallet.py" />
-    <Compile Include="payturetypes\encodedata.py" />
-    <Compile Include="payturetypes\ewallet.py" />
-    <Compile Include="payturetypes\inpay.py" />
-    <Compile Include="payturetypes\merchant.py" />
-    <Compile Include="payturetypes\paytureresponse.py" />
-    <Compile Include="payturetypes\transaction.py" />
-    <Compile Include="payturetypes\__init__.py" />
-    <Compile Include="setup.py">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="testapp\router.py">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="testapp\starttest.py">
-      <SubType>Code</SubType>
-    </Compile>
-  </ItemGroup>
-  <PropertyGroup>
-    <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
-    <PtvsTargetsFile>$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets</PtvsTargetsFile>
-  </PropertyGroup>
-  <Import Condition="Exists($(PtvsTargetsFile))" Project="$(PtvsTargetsFile)" />
-  <Import Condition="!Exists($(PtvsTargetsFile))" Project="$(MSBuildToolsPath)\Microsoft.Common.targets" />
-  <!-- Uncomment the CoreCompile target to enable the Build command in
-       Visual Studio and specify your pre- and post-build commands in
-       the BeforeBuild and AfterBuild targets below. -->
-  <!--<Target Name="CoreCompile" />-->
-  <Target Name="BeforeBuild">
-  </Target>
-  <Target Name="AfterBuild">
-  </Target>
-</Project>
