--- payture/payture/payturetypes/__init__.py
@@ -1,12 +1,0 @@
-import constants 
-import paytureresponse
-import cardinfo
-import encodedata
-import transaction 
-import digitalwallet
-import api
-import ewallet
-import inpay
-import merchant
-
-
