--- payture/payture/testapp/commandslist.txt
@@ -1,28 +1,0 @@
-==================================================================================	
-Commands for help:
-
-	* fields - list current key-value pairs that used in request to Payture server. 
-	* changefields - command for changing current values of  key-value pairs that used in request to Payture server.  
-	* commands - list avaliable commands for this console program.  
-	* changemerchant - commands for changing current merchant account settings. 
-	* help - commands that types this text (description of commands that you can use in this console program.)
-
-==================================================================================	
-Commands for invoke PaytureAPI functions:
-	* pay - use for one-stage payment. In EWALLET an INPAY api this command can be use for block funds - if you specify SessionType=Block. 
-	* block - use for block funds on Customer card. After that command the funds can be charged by Charge command or unblocked by Unblock command. This command use only for API.  
-	* charge - write-off of funds from customer card.  
-	* unblock - unlocking of funds on customer card. 
-	* refund - operation for refunds.  
-	* getsstate - use for getting the actual state of payments in Payture processing system. This command use only for API. 
-	* paystatus - use for getting the actual state of payments in Payture processing system. This command use for EWALLET and INPAY.  
-	* init - use for payment initialization, customer will be redirected on Payture payment gateway page for enter card's information.  
-	* register - register new customer. This command use only for EWALLET. 
-	* check - check for existing customer account in Payture system. This command use only for EWALLET.  
-	* update - This command use only for EWALLET.  
-	* delete - delete customer account from Payture system. This command use only for EWALLET. 
-	* add - register new card in Payture system. This command use only for EWALLET. 
-	* activate - activate registered card in Payture system. This command use only for EWALLET.    
-	* sendcode - provide additional authentication for customer payment. This command use only for EWALLET. 
-	* remove - delete card from Payture system. This command use only for EWALLET. 
-==================================================================================
